Django_Cache/pyproject.toml
[project]
name = "python-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-debug-toolbar>=6.3.0",
    "django-filter>=25.2",
    "djangorestframework>=3.17.1",
    "docutils>=0.22.4",
    "drf-spectacular>=0.29.0",
    "flake8>=7.3.0",
    "flake8-docstrings>=1.7.0",
    "pillow>=12.2.0",
]
